+++ src/IW7/resolver.hpp
@@ -16,8 +16,8 @@ class resolver
     static std::unordered_map<opcode, std::string> opcode_map;
     static std::unordered_map<std::uint16_t, std::string> builtin_function_map;
     static std::unordered_map<std::uint16_t, std::string> builtin_method_map;
-    static std::unordered_map<std::uint16_t, std::string> file_map;
-    static std::unordered_map<std::uint16_t, std::string> token_map;
+    static std::unordered_map<std::uint32_t, std::string> file_map;
+    static std::unordered_map<std::uint32_t, std::string> token_map;
 
 public:
     static auto opcode_id(const std::string& name) -> opcode;
@@ -29,11 +29,11 @@ public:
     static auto builtin_method_id(const std::string& name) -> std::uint16_t;
     static auto builtin_method_name(std::uint16_t id) -> std::string;
 
-    static auto file_id(const std::string& name) -> std::uint16_t;
-    static auto file_name(std::uint16_t id) -> std::string;
+    static auto file_id(const std::string& name) -> std::uint32_t;
+    static auto file_name(std::uint32_t id) -> std::string;
 
-    static auto token_id(const std::string& name) -> std::uint16_t;
-    static auto token_name(std::uint16_t id) -> std::string;
+    static auto token_id(const std::string& name) -> std::uint32_t;
+    static auto token_name(std::uint32_t id) -> std::string;
 
     static auto find_builtin_func(const std::string& name) -> bool;
     static auto find_builtin_meth(const std::string& name) -> bool;
